create centipede class
